11.19.24
Was messing around with ordinal code that we did in class and trying to figure out if it's easier to implement the cipher with or without the ord() function. I'm honestly still a little confused on how I could properly implement this into my program so I've been doing some researching. Just trying to take my time with it and understand the process without rushing it.